Make multiple inclusing possible
If you're working on a project using the littleWire.h multiple times,
you run into linking problems (duplicate symbols). Therefore all defined
variables should be defined in the .cpp file (one place) and should not
be in the header file. If they're in the header file, they are defined
in every file that includes the header.